app/workers.py
from celery import Celery
import os
from dotenv import load_dotenv

load_dotenv()


CELERY_BROKER_URL = os.getenv("REDIS_URL", "redis://redis:6379/0")
CELERY_RESULT_BACKEND = os.getenv("REDIS_URL", "redis://redis:6379/0")

celery_app = Celery(
    "worker",
    broker=CELERY_BROKER_URL,
    backend=CELERY_RESULT_BACKEND
)


celery_app.autodiscover_tasks(['app.tasks'])