requirements.txt
fastapi
uvicorn[standard]
celery
redis
sqlalchemy
psycopg2-binary
python-dotenv
docker==7.1.0
requests==2.31.0